feat(models): 完成三阶段模型一致性修复
Phase 1-3 全部任务完成：
- ID类型统一（跨模块）
- 评分体系统一（Bookstore）
- 金额字段统一（跨模块）
- 进度口径统一（Reader）
- 角色值统一（Auth & Users）
- JSON/BSON命名统一（跨模块）
- 审计字段统一（Audit）

修改范围：
- 70+ 模型文件
- 40+ Repository方法
- 30+ Service方法
- 新增 types 包（7个类型定义文件）
- 26个文件的JSON/BSON标签标准化

测试验证：
- 编译通过 ✅
- 核心测试通过（60+测试用例）✅
- 类型基础设施正常工作 ✅

相关文档：
- docs/plans/2026-01-22-model-consistency-fix-three-phase-design.md
- docs/2026-01-22-model-consistency-fix-compilation-verification-report.md
- docs/2026-01-22-model-consistency-fix-test-verification-report.md

+++ service/finance/wallet/wallet_service.go
@@ -79,7 +79,7 @@ func (s *WalletServiceImpl) GetBalance(ctx context.Context, userID string) (int6
 		return 0, fmt.Errorf("钱包不存在")
 	}
 
-	return wallet.Balance, nil
+	return int64(wallet.Balance), nil
 }
 
 // FreezeWallet 冻结钱包（根据用户ID）
@@ -135,7 +135,7 @@ func convertToWalletResponse(wallet *financeModel.Wallet) *Wallet {
 	return &Wallet{
 		ID:        wallet.ID,
 		UserID:    wallet.UserID,
-		Balance:   wallet.Balance,
+		Balance:   int64(wallet.Balance),
 		Frozen:    wallet.Frozen,
 		CreatedAt: wallet.CreatedAt,
 		UpdatedAt: wallet.UpdatedAt,
